pyproject.toml
[tool.poetry]
name = "lifemed_api"
version = "0.1.0"
description = "API RESTful for management of medical clinics. Stack: Python, FastAPI and PostgreSQL"
authors = ["Lucas Saporetti <lucassaporetti@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pydantic = "^1.8.2"
uvicorn = "0.14.0"
fastapi = "0.68.0"
SQLAlchemy = "==1.4.22"
alembic = "^1.6.5"
gunicorn = "20.1.0"
psycopg2-binary = "^2.9.1"
python-json-logger = "^2.0.2"
PyJWT = {extras = ["crypto"], version = "^2.8.0"}
cryptography = "^42.0.7"


[tool.poetry.dev-dependencies]
pylint = "^2.9.6"
flake8 = "==3.9.2"
black = "==21.7b0"
isort = "==5.9.3"
bandit = "^1.7.0"
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
requests-mock = "^1.9.3"
pytest-cov = "^2.12.1"
safety = "^1.10.3"
Sphinx = "^4.1.2"
recommonmark = "^0.7.1"
sphinx-rtd-theme = "^0.5.2"
aiofiles = "^0.7.0"
mock = "^4.0.3"
ipython = "^7.26.0"
ipdb = "^0.13.9"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "0.14.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"